Remove references to misc_kernels.h
There are a few files which still reference `misc_kernels.h`, which was
deleted in an earlier commit c7325a5267b58fc3391f5b912e35ebfc0723d74e.
This is a problem when trying to build the entire repository with a glob
that just targets all the `.cu` / `.h` / `.hpp` files.